Refactor  search/browser.rs to browser_rows.rs
