+++ rslib/backend.proto
@@ -103,6 +103,7 @@ enum ServiceIndex {
   SERVICE_INDEX_I18N = 12;
   SERVICE_INDEX_COLLECTION = 13;
   SERVICE_INDEX_CARDS = 14;
+  SERVICE_INDEX_BROWSER_ROWS = 15;
 }
 
 service SchedulingService {
@@ -234,7 +235,6 @@ service SearchService {
   rpc JoinSearchNodes(JoinSearchNodesIn) returns (String);
   rpc ReplaceSearchNode(ReplaceSearchNodeIn) returns (String);
   rpc FindAndReplace(FindAndReplaceIn) returns (OpChangesWithCount);
-  rpc BrowserRowForCard(CardID) returns (BrowserRow);
 }
 
 service StatsService {
@@ -275,6 +275,10 @@ service CardsService {
   rpc RemoveCards(RemoveCardsIn) returns (Empty);
   rpc SetDeck(SetDeckIn) returns (OpChanges);
   rpc SetFlag(SetFlagIn) returns (OpChanges);
+}
+
+service BrowserRowsService {
+  rpc BrowserRowForCard(CardID) returns (BrowserRow);
 }
 
 // Protobuf stored in .anki2 files
