+++ rslib/src/search/mod.rs
@@ -1,7 +1,6 @@
 // Copyright: Ankitects Pty Ltd and contributors
 // License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 
-pub mod browser;
 mod cards;
 mod notes;
 mod parser;
